(#291) LSPMain: refactor into PowerShellSettings (app service) and LanguageServer (project service)
