+++ src/main/kotlin/com/intellij/plugin/powershell/lang/lsp/ide/listeners/EditorLSPListener.kt
@@ -5,15 +5,15 @@ package com.intellij.plugin.powershell.lang.lsp.ide.listeners
 
 import com.intellij.openapi.editor.event.EditorFactoryEvent
 import com.intellij.openapi.editor.event.EditorFactoryListener
-import com.intellij.plugin.powershell.lang.lsp.LSPInitMain
+import com.intellij.plugin.powershell.lang.lsp.LanguageServer
 
 class EditorLSPListener : EditorFactoryListener {
 
   override fun editorReleased(event: EditorFactoryEvent) {
-    LSPInitMain.editorClosed(event.editor)
+    LanguageServer.editorClosed(event.editor)
   }
 
   override fun editorCreated(event: EditorFactoryEvent) {
-    LSPInitMain.editorOpened(event.editor)
+    LanguageServer.editorOpened(event.editor)
   }
 }
