+++ src/main/resources/META-INF/plugin.xml
@@ -110,8 +110,6 @@
                              bundle="messages.MessagesBundle"
                              key="settings.powershell"/>
 
-    <applicationService serviceImplementation="com.intellij.plugin.powershell.lang.lsp.LSPInitMain"/>
-
     <editorFactoryListener implementation="com.intellij.plugin.powershell.lang.lsp.ide.listeners.EditorLSPListener"/>
 
     <errorHandler implementation="com.intellij.diagnostic.JetBrainsMarketplaceErrorReportSubmitter"/>
